pyproject.toml
[tool.poetry]
name = "event-planner"
version = "0.1.0"
description = ""
authors = ["Juliana Zacharias <julianazacharias.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "event_planner"}]

[tool.poetry.dependencies]
python = "^3.10"
isort = "^5.12.0"
blue = "^0.9.1"
pip-audit = "^2.6.1"
mkdocs = "^1.5.2"
pytest = "^7.4.1"
pytest-coverage = "^0.0"
flask = "^2.3.3"
flask-sqlalchemy = "^3.0.5"
flask-jwt-extended = "^4.5.2"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
flask-restx = "^1.1.0"
flask-uploads = "^0.2.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
